5363 - Add skeleton
