+++ src/client/pages/Login/ChangePassword/ChangePasswordForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ChangePasswordForm'
