+++ src/client/pages/Login/ChangePassword/ChangePasswordSkeleton/ChangePasswordSkeleton.scss
@@ -1,0 +1,6 @@
+@import 'src/client/style/partials';
+
+.change-password-skeleton {
+  display: grid;
+  grid-row-gap: 40px;
+}
