+++ src/client/pages/Login/ChangePassword/ChangePasswordSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ChangePasswordSkeleton'
